place useGetItemById in useTree
