+++ packages/fluentui/react-northstar/src/components/Tree/hooks/useTree.tsx
@@ -1,8 +1,7 @@
 import * as React from 'react';
 import { ObjectShorthandCollection } from '../../../types';
 import { TreeItemProps } from '../TreeItem';
-import { flattenTree } from './flattenTree';
-import { useGetItemById } from './useGetItemById';
+import { BaseFlatTree, BaseFlatTreeItem, flattenTree } from './flattenTree';
 import { useTreeActiveState } from './useTreeActiveState';
 import { useTreeBehavior } from './useTreeBehavior';
 
@@ -15,6 +14,26 @@ export interface UseTreeOptions {
   defaultActiveItemIds?: string[];
   /** Only allow one subtree to be expanded at a time. */
   exclusive?: boolean;
+}
+
+type GetItemById = (id: string) => BaseFlatTreeItem;
+
+/**
+ * This hook returns a stable `getItemById()` function that will lookup in latest `flatTree`.
+ * This is used used to have stable callbacks that can be passed to React's Context.
+ */
+function useGetItemById(flatTree: BaseFlatTree): GetItemById {
+  // An exception is thrown there to ensure that a proper callback will assigned to ref
+  const callbackRef = React.useRef<GetItemById>(() => {
+    throw new Error('Callback is not assigned yet');
+  });
+  // We are assigning a callback during render as it can be used during render and in event handlers
+  callbackRef.current = itemId => {
+    return flatTree[itemId];
+  };
+  return React.useCallback<GetItemById>((...args) => {
+    return callbackRef.current(...args);
+  }, []);
 }
 
 export function useTree(props: UseTreeOptions) {
